직급, 역할 recoilState로 관리
