+++ src/states/data/hierarchyListState.ts
@@ -1,0 +1,14 @@
+import { atom } from "recoil";
+
+export type Hierarchy = {
+  id: number;
+  nameKr: string;
+  nameEng: string;
+};
+
+const hierarchyListState = atom<Hierarchy[]>({
+  key: "hierarchyListState",
+  default: [],
+});
+
+export default hierarchyListState;
